Added frontend UI

+++ backend/controllers/goalController.js
@@ -1,15 +1,16 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find({})
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
-
+ 
 // @desc    POST goals
 // @route   POST /api/goals
 // @access  Private
@@ -19,7 +20,11 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add a text field')
     }
 
-    const goal = await Goal.create({text: req.body.text})
+    const goal = await Goal.create({
+        text: req.body.text,
+        user: req.user.id,
+    })
+
     res.status(200).json(goal)
 })
 
@@ -32,6 +37,19 @@ const updateGoal = asyncHandler(async (req, res) => {
     if (!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true,})
@@ -48,6 +66,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     
     await goal.deleteOne()
